backend/requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
Flask==3.0.2
Flask-Cors==4.0.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pdfminer.six==20231228
pycparser==2.21
rfc3986==1.5.0
sniffio==1.3.0
Werkzeug==3.0.1
